src/oai-proxy/requirements.txt
openai==0.27.8
requests==2.31.0
flask==2.3.2
python-dotenv==1.0.0
google-cloud-secret-manager==2.16.3